refactor: split db/activities.ts into merge/supersession/mutations/queries (#585)
Pure refactor — no logic changes. The 1216-line `db/activities.ts` is replaced
by a `db/activities/` directory split by responsibility:

- merge.ts (367) — Pure cross-source + same-type merge logic, no DB access
- mutations.ts (369) — Insert/update/delete writes; triggers supersession
- queries.ts (366) — Read operations
- supersession.ts (141) — `superseded_by` materialization
- index.ts (50) — Barrel re-export

Consumer imports of `./activities.ts` updated to `./activities/index.ts` in
db/index.ts, db/health-connect.ts, and the two integration test files.
